requirements.txt
alphashape==1.3.1
astropy==6.1.0
matplotlib==3.8.0
networkx==3.1
numpy==2.0.0
opencv_contrib_python==4.8.1.78
opencv_python==4.8.1.78
openslide_python==1.3.1
pandas==2.2.2
scikit_learn==1.3.1
scipy==1.13.1
seaborn==0.13.2
Shapely==2.0.4
skimage==0.0
tqdm==4.66.1
valis_wsi==1.0.0rc12
